1. Completed: Merging the latest changes from Crhistian for the PR: Fix sale pricing bugs found by Ashley #363.
1. Completed: Merging the latest changes from Crhistian for the PR: Fix sale pricing bugs found by Ashley #363.

+++ src/Middleware/src/Headstart.API/Commands/LocationPermissionCommand.cs
@@ -233,7 +233,7 @@ namespace Headstart.API.Commands
 			try
 			{
 				var hasAccess = await IsUserInAccessGroup(locationId, UserGroupSuffix.PermissionAdmin.ToString(), decodedToken);
-				Require.That(hasAccess, new ErrorCode(@"Insufficient Access", $@"The User cannot manage permissions for: {locationId}.", (int)HttpStatusCode.Forbidden));
+				Require.That(hasAccess, new ErrorCode(@"Insufficient Access", $@"The User cannot manage permissions for: {locationId}.", HttpStatusCode.Forbidden));
 			}
 			catch (Exception ex)
 			{
