+++ src/Middleware/src/Headstart.API/Commands/SupplierSync/SupplierSyncCommand.cs
@@ -61,7 +61,7 @@ namespace Headstart.API.Commands.SupplierSync
 				var supplieId = id.Split("-").Length > 1 ? id.Split("-")[1] : id;
 				if (orderType != OrderType.Quote)
 				{
-					Require.That(decodedToken.CommerceRole == CommerceRole.Seller || supplieId == me.Supplier.ID, new ErrorCode(@"Unauthorized", @"You are not authorized view this order.", (int)HttpStatusCode.Unauthorized));
+					Require.That(decodedToken.CommerceRole == CommerceRole.Seller || supplieId == me.Supplier.ID, new ErrorCode(@"Unauthorized", @"You are not authorized view this order.", HttpStatusCode.Unauthorized));
 				}
 				try
 				{
